requirements.txt
flask
mysql-connector-python
psutil
requests
pandas
scikit-learn
gunicorn
flask-cors
